backend/requirements.txt
# CareerCompass Python Dependencies
# Made by stealthcoderX | All rights reserved.

Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
PyMySQL==1.1.1
Werkzeug==3.0.3
python-dotenv==1.0.1
cryptography==42.0.8
gunicorn==22.0.0
